scripts/requirements.txt
PyEmail==0.0.1
pymysql==1.0.2
PyExecJS==1.5.1
